Export ENV vars via dokku config
Previously the ENV vars were provided to the application by the docker
`-e` option. While this works, it is confusing since dokku provides a
config vars management and those variables don’t show up in it.

To address the issue, `dokku config:set-norestart` and `dokku
config:unset-norestart` are used when creating a new database, deleting
one or starting the Postgresql container.
When the Postgresql container is started, ENV vars are exported again
for every application using a Postgresql DB.

Since the connection configuration is now available in a normalized way
(like on Heroku), linking an application to another DB is just a matter
of using the desired ENV var. That’s why `link` and `unlink` actions
have been deleted.
Original connection configuration for a given application is always
available with the `dokku psql:url` command, so it’s relatively easy to
revert to the original state.