Extract common code into a functions script
psql:install is no more and install code now lies in the install script.

Auto start at pre-deploy and pre-build has been removed for now since
it will cause a global restart of all applications when deploying or
building an application. It will be reintroduced later with a better
implementation.